+++ examples/tokio.rs
@@ -1,20 +1,20 @@
-#[cfg(feature = "tokio")]
+#[cfg(feature = "use_tokio")]
 extern crate futures;
-#[cfg(feature = "tokio")]
+#[cfg(feature = "use_tokio")]
 extern crate sysfs_gpio;
-#[cfg(feature = "tokio")]
-extern crate tokio_core;
+#[cfg(feature = "use_tokio")]
+extern crate tokio;
 
-#[cfg(feature = "tokio")]
-use futures::{Future, Stream};
-#[cfg(feature = "tokio")]
-use sysfs_gpio::{Direction, Edge, Pin};
-#[cfg(feature = "tokio")]
+#[cfg(feature = "use_tokio")]
 use std::env;
-#[cfg(feature = "tokio")]
-use tokio_core::reactor::Core;
 
-#[cfg(feature = "tokio")]
+#[cfg(feature = "use_tokio")]
+use futures::{Future, lazy, Stream};
+
+#[cfg(feature = "use_tokio")]
+use sysfs_gpio::{Direction, Edge, Pin};
+
+#[cfg(feature = "use_tokio")]
 fn stream(pin_nums: Vec<u64>) -> sysfs_gpio::Result<()> {
     // NOTE: this currently runs forever and as such if
     // the app is stopped (Ctrl-C), no cleanup will happen
@@ -22,26 +22,26 @@ fn stream(pin_nums: Vec<u64>) -> sysfs_gpio::Result<()> {
     // can be done about this as Rust signal handling isn't
     // really present at the moment.  Revisit later.
     let pins: Vec<_> = pin_nums.iter().map(|&p| (p, Pin::new(p))).collect();
-    let mut l = Core::new()?;
-    let handle = l.handle();
-    for &(i, ref pin) in pins.iter() {
-        pin.export()?;
-        pin.set_direction(Direction::In)?;
-        pin.set_edge(Edge::BothEdges)?;
-        handle.spawn(pin.get_value_stream(&handle)?
-                         .for_each(move |val| {
-                                       println!("Pin {} changed value to {}", i, val);
-                                       Ok(())
-                                   })
-                         .map_err(|_| ()));
-    }
-    // Wait forever for events
-    loop {
-        l.turn(None)
-    }
+    let task = lazy(move || {
+        for &(i, ref pin) in pins.iter() {
+            pin.export().unwrap();
+            pin.set_direction(Direction::In).unwrap();
+            pin.set_edge(Edge::BothEdges).unwrap();
+            tokio::spawn(pin.get_value_stream().unwrap()
+                .for_each(move |val| {
+                    println!("Pin {} changed value to {}", i, val);
+                    Ok(())
+                })
+                .map_err(|_| ()));
+        }
+        Ok(())
+    });
+    tokio::run(task);
+
+    Ok(())
 }
 
-#[cfg(feature = "tokio")]
+#[cfg(feature = "use_tokio")]
 fn main() {
     let pins: Vec<u64> = env::args()
         .skip(1)
@@ -54,7 +54,7 @@ fn main() {
     }
 }
 
-#[cfg(not(feature = "tokio"))]
+#[cfg(not(feature = "use_tokio"))]
 fn main() {
     println!("This example requires the `tokio` feature to be enabled.");
 }